Updated git submodules
Project: yangtools master 79a740220a9b90d06446eb094c4bf6795cd314bc

BUG-5222: Optimize use of declared substatements

Do not store the return in local variable, as that serves
no real purpose.

Change-Id: Iff1ae8d2d5fe6035c283c559a38fd928637cb3c6
Signed-off-by: Robert Varga <rovarga@cisco.com>


Remove StatementIdentifier

Since we have started identifying children using their declaration
offset, we have no real use for StatementIdentifier. Remove it and
adjust internal users to refer to the raw argument instead.

Change-Id: I7a789935142c8cd63848fb185d6350bfb501e475
Signed-off-by: Robert Varga <rovarga@cisco.com>


Do not confuse statement and argument names

Our extension handling is still flaky, in that it assumes
that implicit extensions take an argument. As we progress
to make these special-cases handled, the confusion is being
flushed out.

This particular error was found when we made argument name
conditional on whether we actually are seeing an argument
being used in the model.

Change-Id: I4b8a0e3aa133343fc65dca38c1dd9e083113fb00
Signed-off-by: Robert Varga <rovarga@cisco.com>


Fix mandatory statement checking

This fixes a mismatch between the comment and what the method
actually does. For some reason we ended up iterating over the same
collection twice.

Furthermore use .forEach() with a self-referencing lambda instead
of Iterables.concat(), as that can be optimized by the implementation.

Change-Id: I6f74794fd58576b9ce34b6f27217b5e81bef4457
Signed-off-by: Robert Varga <rovarga@cisco.com>