Updated git submodules
Project: faas master 59e3b63d3da0deb906391574080c14914dabbfb2

Merge "Do not wrap Guava as a bundle in features' definition"

Do not wrap Guava as a bundle in features' definition

As it will refresh the bundle when the feature is loaded,
because it will certainly be already installed, and hence
will refresh all bundles depending on Guava.

Also, this is not necessary.

This patch is not directly tied to BUG-6956 but it's a result of it

Change-Id: I7a48a54cacd2667515600ee5fa2669f3c12da74d
Signed-off-by: Alexis de Talhouët <adetalhouet@inocybe.com>