Updated git submodules
Project: lispflowmapping master 58c9748104af484a5816095fda6385a153c880ac

Bug 7586: Make Extended SingleFeatureTest pass

We missed the fact that southbound is a dependency for mappingservice,
because of the RPCs to send reply packets. This patch corrects the
feature dependency and allows the new extended SingleFeatureTest to
pass, so we can be removed from the feature blacklist.

Change-Id: I5aadb06cdb2c9ce1bd9c1c3b4efdade53e4a1bcc
Signed-off-by: Lorand Jakab <lojakab@cisco.com>