Updated git submodules
Project: sfc master 2ca1e65cd7c97df60b071ccdc6c49ac216d62899

Act on sonar reported issues / technical debt

Stop using the 'order' local variable in the SfcOfFlowProgrammerImpl
methods; since we're only appending the action to the action list, I
think it is more explicit / less error-prone to indicate the 'order' as
the current-list size (always append to the end of the list)

The configureNshNscTransportEgressFlow in the SfcOfFlowProgrammerImpl
class had a bug; the next protocol was not deing correctly defined as
NSH - as a result, that action list only featured a single action, which
is to set the output port. Now it has an action setting the next
protocol to NSH, and another setting the output protocol.

Tests are updated accordingly.

Used a StringBuilder instead of a StringBuffer in the SfcOfRspProcessor
because the StringBuffer is thread safe, which, in this scenario only
leads to throttled performance.

Removed the TUN_GPE_NP_NSH constant from yet another class.

Change-Id: Ie66ba64f4ec015e97856da5289c19f10efdf3566
Signed-off-by: Miguel Duarte <miguel.duarte.de.mora.barroso@ericsson.com>