Updated git submodules
Project: lispflowmapping master 9a100b81e9f2875873d7e13482abbe1160542020

Add mapping record with port data add with fixed ip

When a neutron port create comes with a fixed ip, we consider the fixed
ip as EID. We need to save this EIDs specific to host id, so that when
we have host id to locator mapping available we can create a mapping
record.

This patch adds hostid to all eid mappings keyed by port uuid. And also
attempts to create a mapping record if hostid to locator mapping is
available.

Change-Id: Ia6ae4643e1ecf5d16c592bcd9a52170b28f38777
Signed-off-by: Shakib Ahmed <sheikahm@cisco.com>