Updated git submodules
Project: netvirt stable/carbon 57b304e56486041b16c68dfe5285ce3743cf1f55

Bug-7718 Operational neutron port status

networking-odl requires feedback as to when a neutron
port can be marked as "ACTIVE". This commit places an
estimation of that information in the operational data
store. For ports that connect to OVS we wait for them
to connect and for basic L2 flows to be configured.
Note that we do not validate the flows in operational
nor do we check all flows. Ports that are implemented
purely as flows are marked ACTIVE immediately. The plan
is for networking-odl to receive notifications of these
status changes via a websocket.

It was decided at the time being
to keep the port status out of the ODL neutron project
since (a) functionally, this issue is mainly an issue for
netvirt and (b) to do it right in netvirt would require
some re-architecting.

Change-Id: Id719e904b277fe4dbb9c3d118d24c3bedf110a33
Signed-off-by: Josh <jhershbe@redhat.com>