Updated git submodules
Project: netvirt master c8eed19fb019b282ed05158f316f19726989aa2c

Bug 7922 - Use counter to keep track of duplicate flow entries

Problem: Single flow entry associated with two more security rules
incorrectly deleted when one of the security rule is deleted.

Fix: Use counter to keeps track of duplicate security rules
in a particular VM. The counter is maintained in the flow cookie.
When writing a security flow, the flow counter is incremented
if the flow already exists, otherwise the flow cookie is set to '1'.
When removing a security flow, if the counter value is '1',
the flow is removed from the node, otherwise the flow counter is
decremented.

This changeset also:
- reversed some previous partial fixes for the same problem.
- refactor common Acl Service to AbstractAclService
- reformat/clean up code

Change-Id: Idfec36fa47c88db76a8382f8fcbc65c4100e9e6e
Signed-off-by: Vinh Nguyen <vinh.nguyen@hcl.com>