Updated git submodules
Project: lispflowmapping master 35296ebb7ae0aa22cd2ee4938941f17b0d54956c

Bug 7818: xTR-ID timestamp updated only when merging

A warning was shown on the Map-Register fast path with default setting,
saying somthing "Could not update timestamp for EID 192.0.2.0/24 xTR-ID
E073DF027BF20605DC9B9518BA753345, no mapping found". Default settings
mean that mapping merge is not on and xTR-ID mappings are not stored, so
the warning is not warranted. And it's not just the warning, the extra
lookup at EVERY fast path Map-Register timestamp upadte has a
performance cost, so make sure it only happens when merge is on.

Change-Id: Id7a66ff8c3309bbcb4ef0bf014f2987149f02fe7
Signed-off-by: Lorand Jakab <lojakab@cisco.com>