Updated git submodules
Project: netconf master 341160de933eda6c0a48aa44196315ebac0dfcaf

Merge "Bug 6645 - 2 digits milliseconds can not be parsed in notification eventTime"

Bug 6645 - 2 digits milliseconds can not be parsed in notification eventTime

Create a Parser and a Formater based of Java 8 DateTimeFormater API, using
the ISO_DATE_TIME format, the UTC time-zone offset, and the local time-zone.

As we're moving away from the pattern approach to parse/format those values,
we have to handle the leap second corner case; it is possible to encounter
time value being 23:59:60, see https://en.wikipedia.org/wiki/Leap_second.
For that reason, if the parsing fails, we catch the exception and check whether
the time has a leap second, and if it has, we replace the second-of-minute of
60 with 59.

Change-Id: I8fcc66f5a18f1781bf14f895c23cd2b890b5b5a7
Signed-off-by: Alexis de Talhouët <adetalhouet@inocybe.com>