Updated git submodules
Project: netvirt stable/boron 987db2384925c49681315146457ab9c10b2c690a

Bug 7876 : After router association to L3vpn, one of the VM ip is not
removed from router interface to BGPVPN

When a VPN associated to router is deleted and a same VPN is associated
back to same router , there is race condition , where interface is not found
in VpnInterfaceOp and remove event is being ignored.
Hence interface remains in router VPN.

It is also seen that , one of interface from same subnet is swapped and others are
not. Its due to race condition that exist in NeutronVpnManager where
create and update can race with each other. Currently create uses DJC ,
where as update doesn't.

Change-Id: I354c3ce93f965580c735f1153a043e2514d68af7
Signed-off-by: ehvkand <hanamantagoud.v.kandagal@ericsson.com>