Updated git submodules
Project: vtn stable/boron b1971f8c400e27724712a1ab5be8786bbfa596a3

Fix Bug 7969, to avoid incompatible database queries contructed

Postgres used to auto-detect the types of values in the where clause
and the arguments passed to  it. This has been disabled  in the recent
releases. This Fix will avoid a situation where VTN Coordinator
tries to construct  a query with such a problem causing a failure in
postgres and eventually leading to multiple failures.

Change-Id: Ic444a99fc1d2c51cc758f8dba745affb3a0b57d0
Signed-off-by: gvrangan <venkatrangang@hcl.com>