Updated git submodules
Project: netvirt stable/boron 770c08d6d48e59d9a2a4da3fc2b2407d36f4fbab

Bug 7939:  VpnService Suite and Tempest failures

This fix is an add-on patch for the below review

https://git.opendaylight.org/gerrit/#/c/53772/

and is in a direction to solve the OperTxn gets
failing when configTxn changes are triggered.

Also we fix here the routerInterface VrfEntry
management such that when we clearly know its a
router-interface VRFEntry, we donot do any further
processing than what is required for the same.

Change-Id: I8efc5c41f26e9e6dd04c8638b2cfaa94a2050384
Signed-off-by: Vivekanandan Narasimhan <n.vivekanandan@ericsson.com>