Updated git submodules
Project: mdsal stable/boron eba6505fa7ca734ff895fcf753664f3d73f1324b

BUG-8004: handle implicit RPC input

RPC input is always defined implicitly, which is something we cannot
deal with in Binding Specification compatibly, as original generators
did not emit classes for such methods and implicitly map them to Void.

In order to deal with these, we need to recognize when an RPC input
is not declared and re-wire it to a null codec. This loses information,
but it is the best we can do (simply because there is no binding
representation).

Change-Id: I4edbcd0cc886396bdba79f4c0ccfd91ba3d6b2c7
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
(cherry picked from commit a69ca3a80b53cf7697b8a29029260209dea842ac)